fix delete kiemke

+++ Absoft/Repositories/Implimentations/KiemKeChiTietRepository.cs
@@ -24,39 +24,65 @@ namespace Absoft.Repositories.Implimentations
         }
         public async Task<bool> DeleteKiemKeChiTietAsync(int maPKK, int maPN, int maVT, int maKho)
         {
+
             // remove kiemkechitiet
             var kkct = await db.KiemKeChiTiets.FirstOrDefaultAsync(x => x.MaPhieuNhap == maPN && x.MaPhieuKiemKe == maPKK && x.MaVatTu == maVT);
             int sltd = kkct.SoLuongTheoDoi.Value;
             int sltt = kkct.SoLuongThucTon.Value;
-            db.KiemKeChiTiets.Remove(kkct);
+            int slkk = sltt - sltd;
             //tru soluongtong trong ptl
             var kkvt = await db.KiemKeVatTus.FindAsync(maPKK);
             kkvt.TongTheoDoi -= sltd;
             kkvt.TongThucTon -= sltt;
+            kkvt.TongKiemKe = kkvt.TongThucTon - kkvt.TongTheoDoi;
             // tra lai so luong theo doi vao kho
             var kh = await db.KhoHangs.FirstOrDefaultAsync(x => x.MaKho == maKho && x.MaPhieuNhap == maPN && x.MaVatTu == maVT);
-            kh.SoLuongTon = sltd;
+            if (kh.SoLuongTon < slkk) return false;
+
+            kh.SoLuongTon -= slkk;
+
+            db.KiemKeVatTus.Update(kkvt);
+            db.KhoHangs.Update(kh);
+            db.KiemKeChiTiets.Remove(kkct);
+
+
             return await db.SaveChangesAsync() > 0;
+
+
         }
         public async Task<bool> InserKiemKeChiTietAsync(KiemKeChiTietViewModel model, int maPKK, int maKho)
+        {
+
+            // insert chi tiet, update tong kho, update tongsl
+            // var pkk = await db.KiemKeVatTus.FindAsync(maPKK);            
+            var entity = mp.Map<KiemKeChiTiet>(model);
+            entity.MaPhieuKiemKe = maPKK;
+            entity.SoLuongKiemKe = entity.SoLuongThucTon - entity.SoLuongTheoDoi;
+
+            await db.KiemKeChiTiets.AddAsync(entity);
+            await SumSL(model, maPKK);
+            var kh = await db.KhoHangs.FirstOrDefaultAsync(x => x.MaKho == maKho && x.MaPhieuNhap == model.MaPhieuNhap && x.MaVatTu == model.MaVatTu);
+            kh.SoLuongTon += entity.SoLuongKiemKe;
+
+            db.KhoHangs.Update(kh);
+
+            return await db.SaveChangesAsync() > 0;
+
+
+        }
+        public async Task<bool> UpdateKiemKeChiTietAsync(KiemKeChiTietViewModel model, int maPKK, int maKho)
         {
             using (var transaction = db.Database.BeginTransaction())
             {
                 try
                 {
-                    // insert chi tiet, update tong kho, update tongsl
-                    // var pkk = await db.KiemKeVatTus.FindAsync(maPKK);            
-                    var entity = mp.Map<KiemKeChiTiet>(model);
-                    entity.MaPhieuKiemKe = maPKK;
-                    entity.SoLuongKiemKe = entity.SoLuongThucTon - entity.SoLuongTheoDoi;
-                    
-                    await db.KiemKeChiTiets.AddAsync(entity);
-                    await SumSL(model, maPKK);
-                    var kh = await db.KhoHangs.FirstOrDefaultAsync(x => x.MaKho == maKho && x.MaPhieuNhap == model.MaPhieuNhap && x.MaVatTu == model.MaVatTu);
-                    kh.SoLuongTon = model.SoLuongThucTon;
-
+                    // delete 
+                    var rsDelete = await DeleteKiemKeChiTietAsync(maPKK, model.MaPhieuNhap, model.MaVatTu, maKho);
+                    //insert                    
+                    var rsInsert = await InserKiemKeChiTietAsync(model, maPKK, maKho);
+                    if (rsDelete == false || rsInsert == false) return false;
                     transaction.Commit();
-                    return await db.SaveChangesAsync() > 0;                   
+                    return true;
                 }
                 catch (Exception ex)
                 {
@@ -64,35 +90,15 @@ namespace Absoft.Repositories.Implimentations
                 }
             }
         }
-        public async Task<bool> UpdateKiemKeChiTietAsync(KiemKeChiTietViewModel model, int maPKK, int maKho)
-        {
-            var kkvt = await db.KiemKeVatTus.FindAsync(maPKK);
-            
-            var kkct = await db.KiemKeChiTiets.FirstOrDefaultAsync(x=>x.MaPhieuKiemKe == maPKK && x.MaPhieuNhap == model.MaPhieuNhap && x. MaVatTu == model.MaVatTu );
-            kkct.SoLuongTheoDoi = kkct.SoLuongTheoDoi;
-            kkct.SoLuongThucTon = model.SoLuongThucTon;
-            kkct.SoLuongKiemKe = model.SoLuongThucTon - model.SoLuongTheoDoi;
-
-            kkct.GhiChu = model.GhiChu;
-            var kh = await db.KhoHangs.FirstOrDefaultAsync(x => x.MaKho == maKho && x.MaPhieuNhap == model.MaPhieuNhap && x.MaVatTu == model.MaVatTu);
-            kh.SoLuongTon = model.SoLuongThucTon;
-
-            kkvt.TongTheoDoi += model.SoLuongTheoDoi;
-            kkvt.TongTheoDoi -= kkct.SoLuongTheoDoi;
-            kkvt.TongThucTon += model.SoLuongThucTon;
-            kkvt.TongThucTon -= kkct.SoLuongThucTon;
-
-            return await db.SaveChangesAsync() > 0;
-        }
         public async Task<int> GetSoLuongTheoDoi(int maKho, int maPN, int maVT)
         {
-            var sl = (await db.KhoHangs.FirstOrDefaultAsync(x=>x.MaKho == maKho && x.MaPhieuNhap== maPN && x. MaVatTu== maVT)).SoLuongTon.Value;
+            var sl = (await db.KhoHangs.FirstOrDefaultAsync(x => x.MaKho == maKho && x.MaPhieuNhap == maPN && x.MaVatTu == maVT)).SoLuongTon.Value;
             return sl;
         }
 
         public async Task<bool> CheckExistChiTiet(int maPKK)
         {
-            var entity = await db.KiemKeChiTiets.FirstOrDefaultAsync(x=>x.MaPhieuKiemKe == maPKK);
+            var entity = await db.KiemKeChiTiets.FirstOrDefaultAsync(x => x.MaPhieuKiemKe == maPKK);
             if (entity != null) return false;
             else return true;
         }
@@ -101,6 +107,7 @@ namespace Absoft.Repositories.Implimentations
             var entity = await db.KiemKeVatTus.FindAsync(maPKK);
             entity.TongTheoDoi += model.SoLuongTheoDoi;
             entity.TongThucTon += model.SoLuongThucTon;
+            entity.TongKiemKe = entity.TongThucTon - entity.TongTheoDoi;
             return await db.SaveChangesAsync() > 0;
         }
 
@@ -109,6 +116,6 @@ namespace Absoft.Repositories.Implimentations
             var entity = await db.KiemKeChiTiets.FirstOrDefaultAsync(x => x.MaPhieuKiemKe == maPKK && x.MaPhieuNhap == maPN && x.MaVatTu == maVT);
             var model = mp.Map<KiemKeChiTietViewModel>(entity);
             return model;
-        }     
+        }
     }
 }
